Conflicts:
	io-board/stm32f4/common/modules/io/pv_module_io.c

+++ io-board/stm32f4/common/modules/common/c_common_utils.c
@@ -106,7 +106,7 @@ void c_common_utils_floatToString(float num, char * outbuf, char decplaces) {
  *
  */
 long c_common_utils_getSysTickCount() {
-	return CORE_GetSysTick();
+	return (long) CORE_GetSysTick();
 }
 
 /** \brief Habilita o SysTick.
@@ -117,6 +117,11 @@ void c_common_utils_enSysTick() {
 	CORE_SysTickEn();
 }
 
+//long c_common_utils_micros() {
+//	CORE_SysTickEn();
+//	return (long)(CORE_GetSysTick()/(SystemCoreClock/1000000));
+//}
+
 /** \brief Retorna o valor em milissegundos desde o disparo do SysTick.
  *
  * Caso o Registrador de SysTick não esteja rodando quando a função é chamada, o registrador é ativado
@@ -126,6 +131,8 @@ long c_common_utils_millis() {
 	CORE_SysTickEn();
 	return (long)(CORE_GetSysTick()/(SystemCoreClock/1000));
 }
+
+
 
 /**
   * @}
